pyproject.toml
[project]
name = "gemini-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.22.1",
    "litellm>=1.81.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
    "tavily-python>=0.7.19",
]
